refactor(paysolutions): decompose into GatewayClient/Intent/Webhook/Confirmation services
Behavior-preserving regulated money-inflow gateway. Facade keeps 11-method surface +
ctor (forwardRef intact) + delegates; internally constructs 4 sub-services (6 specs +
module + 6 external consumers untouched). The line-1105 Serializable webhook $tx
(3 JEs: receipt + vat60-reversal + surplus-advance, FIFO, idempotency, contract-close)
moved WHOLE; 5 duplicated gateway-fetch blocks collapsed into GatewayClient.createUiPayment.
+characterization for the 4 previously-untested intent creators. tsc 0, tests green, lint 0.
